+++ CHANGELOG.md
@@ -1,7 +1,27 @@
 # CHANGELOG
 
+### Version 1.1.1 *(2025-02-08)*
+* **Reticle target marker** updates:
+  * Added "Disable in Group dungeons" option.
+  * Added auto disable feature:
+    * auto disable reticle target tracking in solo arenas (e.g. Maelstorm arena);
+    * auto disable reticle target tracking in Infinite archive;
+    * auto disable reticle target tracking in group dungeons, group arenas and trials if player is not in group;
+      * auto re-enable reticle target tracking when players joins group while in group dungeon, group arena or trial.
+  * Little perfomance optimizations for reticle target tracking.
+  * Improved PvP-zones detection.
+  * Slightly ajusted default settings.
+* Bug fixed:
+  * Fixed an issue with context menu of ESO friends list and ignored players list.
+  * Fixed duplicate contacts issue.
+  * Fixed spellchecking in Edit contact dialog.
+* Dependencies:
+  * **[LibExtendedJournal](https://www.esoui.com/downloads/info4031-LibExtendedJournal.html)** no longer included in addon archive and should be installed separately *(it was released as a standalone library)*.
+* Translation:
+  * Updated **German** translation (by **[Baertram](https://www.esoui.com/forums/member.php?u=2028)**).
+
 ### Version 1.1.0 *(2025-02-07)*
-* Added reticle target marker.
+* Added Reticle target marker.
 * Added a watcher to auto-rename contacts when associated player from ESO ingame friends has been renamed.
 * Code updates for **LibCustomMenu** v7.3.0
 * Bug fixes: